+++ tests/archives/grid_archive_test.py
@@ -23,7 +23,7 @@ def assert_archive_elite(archive, solution, objective, measures, grid_indices,
     assert np.isclose(elite.solution, solution).all()
     assert np.isclose(elite.objective, objective).all()
     assert np.isclose(elite.measures, measures).all()
-    assert elite.index == archive.grid_to_int_index(grid_indices)
+    assert elite.index == archive.grid_to_int_index([grid_indices])
     assert elite.metadata == metadata
 
 
@@ -125,7 +125,17 @@ def test_grid_to_int_index(data):
                                            ])[0] == data.int_index)
 
 
+def test_grid_to_int_index_wrong_shape(data):
+    with pytest.raises(ValueError):
+        data.archive.grid_to_int_index([data.grid_indices[:-1]])
+
+
 def test_int_to_grid_index(data):
     assert np.all(
         data.archive.int_to_grid_index([data.int_index])[0] ==
         data.grid_indices)
+
+
+def test_int_to_grid_index_wrong_shape(data):
+    with pytest.raises(ValueError):
+        data.archive.int_to_grid_index(data.int_index)